Move configuration derive to a separate module
